fix wrong int/float conversion
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/gtg/trunk@34 0c38aafd-2db3-40e8-87a3-ec7d7f65c264

+++ inc/pajeColor.h
@@ -168,9 +168,9 @@ const paje_color_t Paje_get_color(gtg_color_t* p_color) {
 	/* todo */
 	paje_color_t res = NULL;
 	asprintf(&res, "%1.1f %1.1f %1.1f", 
-		 GTG_COLOR_GET_RED(p_color->rgb)/256,
-		 GTG_COLOR_GET_GREEN(p_color->rgb)/256,
-		 GTG_COLOR_GET_BLUE(p_color->rgb)/256);
+		 (float)GTG_COLOR_GET_RED(p_color->rgb)/256,
+		 (float)GTG_COLOR_GET_GREEN(p_color->rgb)/256,
+		 (float)GTG_COLOR_GET_BLUE(p_color->rgb)/256);
 	return res;
 }
